requirements.txt
pyrogram==2.0.106
TgCrypto==1.2.3
requests
rich
pyyaml
ffmpeg-python